refactor: remove dormant frontend modules
